+++ src/dataprotection/azext_dataprotection/manual/_params.py
@@ -91,6 +91,13 @@ def load_arguments(self, _):
         c.argument('include_cluster_scope_resources', arg_type=get_three_state_flag(),
                    options_list=['--include-cluster-scope-resources', '--include-cluster-scope'],
                    help="Boolean parameter to decide whether cluster scope resources are included for restore. By default this is taken as true.")
+        c.argument('vaulted_backup_containers', type=str, nargs='+', options_list=["--container-list", "--vaulted-backup-containers"],
+                   help="List of containers to be backed up inside the VaultStore. Use this parameter for DatasourceType AzureBlob.")
+        c.argument('include_all_containers', arg_type=get_three_state_flag(),
+                   help='Switch parameter to include all containers to be backed up inside the VaultStore. Use this parameter for DatasourceType AzureBlob.')
+        c.argument('storage_account_name', type=str, help='Storage account where the Datasource is present. Use this parameter for DatasourceType AzureBlob.')
+        c.argument('storage_account_resource_group', options_list=['--storage-account-resource-group', '--storage-account-rg'], type=str,
+                   help='Storage account resource group name where the Datasource is present. Use this parameter for DatasourceType AzureBlob.')
         c.argument('backup_hook_references',
                    type=namespaced_name_resource_type,
                    options_list=['--backup-hook-references', '--backup-hook-refs'],
@@ -107,6 +114,14 @@ def load_arguments(self, _):
         c.argument('secret_store_uri', type=str, help="specify the secret store uri to use for authentication")
         c.argument('snapshot_resource_group_name', options_list=['--snapshot-resource-group-name', '--snapshot-rg'], type=str, help="Name of the resource group in which the backup snapshots should be stored")
         c.argument('tags', tags_type)
+
+    with self.argument_context('dataprotection backup-instance update') as c:
+        c.argument('backup_instance_name', type=str, help="Backup instance name.")
+        c.argument('resource_group_name', resource_group_name_type)
+        c.argument('vault_name', vault_name_type)
+        c.argument('vaulted_blob_container_list', type=validate_file_or_dict, options_list=['--vaulted-blob-container-list', '--container-blob-list'],
+                   help="Enter the container list to modify a vaulted blob backup. The output for "
+                   "'az dataprotection backup-instance initialize-backupconfig' needs to be provided as input")
 
     with self.argument_context('dataprotection backup-instance update-policy') as c:
         c.argument('backup_instance_name', type=str, help="Backup instance name.")
@@ -280,6 +295,8 @@ def load_arguments(self, _):
         c.argument('from_prefix_pattern', type=str, nargs='+', help="specify the prefix pattern for start range.")
         c.argument('to_prefix_pattern', type=str, nargs='+', help="specify the prefix pattern for end range.")
         c.argument('restore_configuration', type=validate_file_or_dict, help="Restore configuration for restore. Use this parameter to restore with AzureKubernetesService.")
+        c.argument('vaulted_blob_prefix_pattern', options_list=['--vaulted-blob-prefix-pattern', '--vaulted-blob-prefix'],
+                   type=validate_file_or_dict, help="Specify the prefix pattern for vaulted blobs.")
 
     with self.argument_context('dataprotection backup-instance validate-for-restore') as c:
         c.argument('backup_instance_name', options_list=['--backup-instance-name', '--name', '-n'], type=str, help="Backup instance name.")