+++ src/dataprotection/azext_dataprotection/tests/latest/test_dataprotection_backup_instance_restore.py
@@ -6,6 +6,7 @@
 # pylint: disable=line-too-long
 # pylint: disable=unused-import
 
+import unittest
 from knack.log import get_logger
 from azure.cli.testsdk import ScenarioTest
 from azure.cli.testsdk.scenario_tests import AllowLargeResponse
@@ -51,7 +52,7 @@ class BackupInstanceRestoreScenarioTest(ScenarioTest):
         # As a failsafe, ensure restored disk from previous run is deleted
         test.cmd('az disk delete --name "{restoreDiskName}" --resource-group "{rg}" --yes')
 
-        restore_request = test.cmd('az dataprotection backup-instance restore  initialize-for-data-recovery '
+        restore_request = test.cmd('az dataprotection backup-instance restore initialize-for-data-recovery '
                                    '--datasource-type "{dataSourceType}" --restore-location "{restoreLocation}" --source-datastore "{sourceDataStore}" '
                                    '--recovery-point-id "{recoveryPointId}" --target-resource-id "{restoreDiskId}"').get_output_in_json()
         test.kwargs.update({"restoreRequest": restore_request})
@@ -73,7 +74,7 @@ class BackupInstanceRestoreScenarioTest(ScenarioTest):
         track_job_to_completion(test)
 
     @AllowLargeResponse()
-    def test_dataprotection_backup_instance_restore_blob(test):
+    def test_dataprotection_backup_instance_restore_blob_point_in_time(test):
         test.kwargs.update({
             'dataSourceType': 'AzureBlob',
             'sourceDataStore': 'OperationalStore',
@@ -101,10 +102,59 @@ class BackupInstanceRestoreScenarioTest(ScenarioTest):
                                    '--point-in-time "{restorePointInTime}" --backup-instance-id "{backupInstanceId}"').get_output_in_json()
         test.kwargs.update({"restoreRequest": restore_request})
 
-        # Ensure no other jobs running on datasource. Required to avoid operation clashes.
+        # Ensure no other restore jobs running on datasource. Required to avoid operation clashes.
         wait_for_job_exclusivity_on_datasource(test)
 
         test.cmd('az dataprotection backup-instance validate-for-restore -g "{rg}" --vault-name "{vaultName}" -n "{backupInstanceName}" --restore-request-object "{restoreRequest}"')
+        restore_trigger_json = test.cmd('az dataprotection backup-instance restore trigger -g "{rg}" --vault-name "{vaultName}" '
+                                        '-n "{backupInstanceName}" --restore-request-object "{restoreRequest}"').get_output_in_json()
+        test.kwargs.update({"jobId": restore_trigger_json["jobId"]})
+
+        test.cmd('az dataprotection job show --ids "{jobId}"', checks=[
+            test.check('properties.dataSourceName', "{dataSourceName}"),
+            test.exists('properties.extendedInfo.recoveryDestination')
+        ])
+
+    # @unittest.skip('Skipping test until datasources are set back up')
+    @AllowLargeResponse()
+    def test_dataprotection_backup_instance_restore_blob_recovery_point(test):
+        test.kwargs.update({
+            'dataSourceType': 'AzureBlob',
+            'sourceDataStore': 'VaultStore',
+            'backupInstanceName': 'clitestsavltdonotdelete-clitestsavltdonotdelete-65176b0a-64ad-4247-9866-19204633c0d5',
+            'dataSourceName': 'clitestsavltdonotdelete',
+            'dataSourceId': '/subscriptions/38304e13-357e-405e-9e9a-220351dcce8c/resourceGroups/clitest-dpp-rg/providers/Microsoft.Storage/storageAccounts/clitestsavltdonotdelete',
+            'restoreLocation': 'centraluseuap',
+            'containerName': 'container1',
+            'restoreContainerName': 'clitestvsdonotdelete',
+            'targetResourceId': '/subscriptions/38304e13-357e-405e-9e9a-220351dcce8c/resourceGroups/clitest-dpp-rg/providers/Microsoft.Storage/storageAccounts/clitestvsdonotdelete',
+        })
+
+        # Recovery Point Restore - Vaulted Blobs
+        recovery_point = test.cmd('az dataprotection recovery-point list --backup-instance-name "{backupInstanceName}" -g "{rg}" --vault-name "{vaultName}"', checks=[
+            test.greater_than('length([])', 0)
+        ]).get_output_in_json()
+        test.kwargs.update({
+            'recoveryPointId': recovery_point[0]['name']
+        })
+
+        restore_request_vs = test.cmd('az dataprotection backup-instance restore initialize-for-item-recovery '
+                                      '--datasource-type "{dataSourceType}" '
+                                      '--restore-location "{restoreLocation}" '
+                                      '--source-datastore "{sourceDataStore}" '
+                                      '--recovery-point-id "{recoveryPointId}" '
+                                      '--target-resource-id "{targetResourceId}" '
+                                      '--container-list "{containerName}"').get_output_in_json()
+        test.kwargs.update({"restoreRequest": restore_request_vs})
+
+        test.addCleanup(test.cmd, 'az storage container delete --name "{containerName}" --account-name "{restoreContainerName}" --auth-mode "login"')
+
+        # Cleanup failsafe, delete existing container at restore location
+        test.cmd('az storage container delete --name "{containerName}" --account-name "{restoreContainerName}" --auth-mode "login"')
+
+        # Ensure no other restore jobs running on datasource. Required to avoid operation clashes.
+        wait_for_job_exclusivity_on_datasource(test)
+
         restore_trigger_json = test.cmd('az dataprotection backup-instance restore trigger -g "{rg}" --vault-name "{vaultName}" '
                                         '-n "{backupInstanceName}" --restore-request-object "{restoreRequest}"').get_output_in_json()
         test.kwargs.update({"jobId": restore_trigger_json["jobId"]})